commit till 4.22

+++ blog/controllers/blogs.js
@@ -1,20 +1,21 @@
 const { application } = require("express")
 const { response } = require("../app")
 const Blog = require("../models/blog")
+const User = require("../models/user")
 
 const blogsRouter = require("express").Router()
 
-blogsRouter.get("/",(request,response) => {
-    Blog.find({})
-        .then(blogs => {
-            response.json(blogs)
-        })
-        .catch(error => next(error))
+blogsRouter.get("/",async (request,response) => {
+   const blogs = await Blog.find({}).populate("user")
+   response.json(blogs)
 })
+
 
 blogsRouter.post("/",async (request,response) => {
 
     let newBlog = request.body
+
+    const user = request.user
 
     if(!(newBlog.title && newBlog.url)){
         response.status(400).end()
@@ -24,10 +25,19 @@ blogsRouter.post("/",async (request,response) => {
       newBlog = {...newBlog,likes: 0}
     }
    
-    const blog = new Blog(newBlog)
+    const blog = new Blog({
+        title: newBlog.title,
+        author: newBlog.author,
+        url: newBlog.url,
+        likes: newBlog.likes,
+        user: user.id
+    })
 
 
     const result = await blog.save()
+
+    user.blogs = user.blogs.concat(result.id)
+    await user.save()
     response.status(201).json(result)
 
 
@@ -35,8 +45,28 @@ blogsRouter.post("/",async (request,response) => {
 
 blogsRouter.delete("/:id", async (request,response) => {
 
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+
+    const blog = await Blog.findById(request.params.id)
+
+    const user = request.user
+
+    if(blog.user.toString() === user.id.toString()){
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+
+        user.blogs = user.blogs.filter((b) => b.id !== blog.id)
+        await user.save()
+    } else {
+        response.status(400).send({error: "invalid user"})
+    }
+
+})
+
+blogsRouter.get("/:id",async (request,response) => {
+
+   const blog =  await Blog.findById(request.params.id)
+    response.status(201).json(blog)
+
 })
 
 blogsRouter.put("/:id", async (request,response) => {
